+++ lib/widgets/wn_user_profile_card.dart
@@ -4,7 +4,6 @@ import 'package:gap/gap.dart';
 import 'package:sloth/l10n/l10n.dart';
 import 'package:sloth/src/rust/api/metadata.dart';
 import 'package:sloth/theme.dart';
-import 'package:sloth/utils/avatar_color.dart';
 import 'package:sloth/utils/formatting.dart';
 import 'package:sloth/utils/metadata.dart';
 import 'package:sloth/widgets/wn_avatar.dart';
@@ -35,7 +34,7 @@ class WnUserProfileCard extends StatelessWidget {
           pictureUrl: metadata?.picture,
           displayName: displayName,
           size: WnAvatarSize.large,
-          color: avatarColorFromPubkey(userPubkey),
+          color: AvatarColor.fromPubkey(userPubkey),
         ),
         Gap(16.h),
         if (displayName != null)